+++ cvat-core/src/session.ts
@@ -210,8 +210,8 @@ function buildDuplicatedAPI(prototype) {
                         prototype.frames.save,
                     );
                 },
-                async ranges() {
-                    const result = await PluginRegistry.apiWrapper.call(this, prototype.frames.ranges);
+                async cachedChunks() {
+                    const result = await PluginRegistry.apiWrapper.call(this, prototype.frames.cachedChunks);
                     return result;
                 },
                 async preview() {
@@ -329,6 +329,7 @@ export class Job extends Session {
     public readonly taskId: number;
     public readonly dimension: DimensionType;
     public readonly dataChunkType: ChunkType;
+    public readonly dataChunkSize: number;
     public readonly bugTracker: string | null;
     public readonly mode: TaskMode;
     public readonly labels: Label[];
@@ -369,7 +370,7 @@ export class Job extends Session {
         delete: CallableFunction;
         restore: CallableFunction;
         save: CallableFunction;
-        ranges: CallableFunction;
+        cachedChunks: CallableFunction;
         preview: CallableFunction;
         contextImage: CallableFunction;
         search: CallableFunction;
@@ -573,7 +574,7 @@ export class Job extends Session {
             delete: Object.getPrototypeOf(this).frames.delete.bind(this),
             restore: Object.getPrototypeOf(this).frames.restore.bind(this),
             save: Object.getPrototypeOf(this).frames.save.bind(this),
-            ranges: Object.getPrototypeOf(this).frames.ranges.bind(this),
+            cachedChunks: Object.getPrototypeOf(this).frames.cachedChunks.bind(this),
             preview: Object.getPrototypeOf(this).frames.preview.bind(this),
             search: Object.getPrototypeOf(this).frames.search.bind(this),
             contextImage: Object.getPrototypeOf(this).frames.contextImage.bind(this),
@@ -684,7 +685,7 @@ export class Task extends Session {
         delete: CallableFunction;
         restore: CallableFunction;
         save: CallableFunction;
-        ranges: CallableFunction;
+        cachedChunks: CallableFunction;
         preview: CallableFunction;
         contextImage: CallableFunction;
         search: CallableFunction;
@@ -1101,7 +1102,7 @@ export class Task extends Session {
             delete: Object.getPrototypeOf(this).frames.delete.bind(this),
             restore: Object.getPrototypeOf(this).frames.restore.bind(this),
             save: Object.getPrototypeOf(this).frames.save.bind(this),
-            ranges: Object.getPrototypeOf(this).frames.ranges.bind(this),
+            cachedChunks: Object.getPrototypeOf(this).frames.cachedChunks.bind(this),
             preview: Object.getPrototypeOf(this).frames.preview.bind(this),
             contextImage: Object.getPrototypeOf(this).frames.contextImage.bind(this),
             search: Object.getPrototypeOf(this).frames.search.bind(this),